+++ cli/src/pcluster/models/hit/hit_cluster_model.py
@@ -71,7 +71,7 @@ class HITClusterModel(ClusterModel):
             else {}
         )
         try:
-            latest_alinux_ami_id = self._get_latest_alinux_ami_id()
+            latest_alinux2_ami_id = self._get_latest_alinux2_ami_id()
 
             master_network_interfaces = self.build_launch_network_interfaces(
                 network_interfaces_count=int(cluster_section.get_param_value("network_interfaces_count")[0]),
@@ -87,7 +87,7 @@ class HITClusterModel(ClusterModel):
                 InstanceType=master_instance_type,
                 MinCount=1,
                 MaxCount=1,
-                ImageId=latest_alinux_ami_id,
+                ImageId=latest_alinux2_ami_id,
                 CpuOptions=master_cpu_options,
                 NetworkInterfaces=master_network_interfaces,
                 DryRun=True,
@@ -110,7 +110,7 @@ class HITClusterModel(ClusterModel):
                     pcluster_config,
                     compute_resource_section,
                     disable_hyperthreading=disable_hyperthreading,
-                    ami_id=latest_alinux_ami_id,
+                    ami_id=latest_alinux2_ami_id,
                     subnet=compute_subnet,
                     security_groups_ids=security_groups_ids,
                     placement_group=queue_placement_group,