+++ tests/integration-tests/tests/common/utils.py
@@ -21,7 +21,6 @@ from utils import get_instance_info
 LOGGER = logging.getLogger(__name__)
 
 OS_TO_OFFICIAL_AMI_NAME_OWNER_MAP = {
-    "alinux": {"name": "amzn-ami-hvm-*.*.*.*-*-gp2", "owners": ["amazon"]},
     "alinux2": {"name": "amzn2-ami-hvm-*.*.*.*-*-gp2", "owners": ["amazon"]},
     "centos7": {"name": "CentOS 7.*", "owners": ["125523088429"]},
     "centos8": {"name": "CentOS 8.*", "owners": ["125523088429", "247102896272"]},
@@ -33,7 +32,6 @@ OS_TO_OFFICIAL_AMI_NAME_OWNER_MAP = {
 
 # Remarkable AMIs are latest deep learning base AMI and FPGA developer AMI without pcluster infrastructure
 OS_TO_REMARKABLE_AMI_NAME_OWNER_MAP = {
-    "alinux": {"name": "Deep Learning Base AMI (Amazon Linux)*", "owners": ["amazon"]},
     "alinux2": {"name": "Deep Learning Base AMI (Amazon Linux 2)*", "owners": ["amazon"]},
     "centos7": {"name": "FPGA Developer AMI*", "owners": ["679593333241"]},
     "ubuntu1804": {"name": "Deep Learning Base AMI (Ubuntu 18.04)*", "owners": ["amazon"]},
@@ -41,7 +39,6 @@ OS_TO_REMARKABLE_AMI_NAME_OWNER_MAP = {
 
 # Pcluster AMIs are latest ParallelCluster official AMIs that align with cli version
 OS_TO_PCLUSTER_AMI_NAME_OWNER_MAP = {
-    "alinux": {"name": "amzn-hvm-x86_64-*", "owners": ["amazon"]},
     "alinux2": {"name": "amzn2-hvm-*-*", "owners": ["amazon"]},
     "centos7": {"name": "centos7-hvm-x86_64-*", "owners": ["amazon"]},
     "centos8": {"name": "centos8-hvm-x86_64-*", "owners": ["amazon"]},