+++ tests/integration-tests/tests/disable_hyperthreading/test_disable_hyperthreading.py
@@ -22,7 +22,7 @@ from tests.common.utils import fetch_instance_slots
 
 
 # Manually disabled HT
-@pytest.mark.dimensions("sa-east-1", "m4.xlarge", "alinux", "sge")
+@pytest.mark.dimensions("sa-east-1", "m4.xlarge", "centos8", "sge")
 @pytest.mark.dimensions("sa-east-1", "m4.xlarge", "ubuntu1804", "sge")
 # HT disabled via CpuOptions
 @pytest.mark.dimensions("sa-east-1", "c5.xlarge", "alinux2", "sge")