+++ tests/integration-tests/tests/storage/test_ebs.py
@@ -87,7 +87,7 @@ def test_ebs_multiple(scheduler, pcluster_config_reader, clusters_factory):
         _test_ebs_correctly_shared(remote_command_executor, mount_dir, scheduler_commands)
 
 
-@pytest.mark.dimensions("cn-northwest-1", "c4.xlarge", "alinux", "slurm")
+@pytest.mark.dimensions("cn-northwest-1", "c4.xlarge", "alinux2", "slurm")
 @pytest.mark.usefixtures("region", "os", "instance")
 def test_default_ebs(scheduler, pcluster_config_reader, clusters_factory):
     cluster_config = pcluster_config_reader()